#1683 Move BatchPoints into Client
